move areas tags and active areas selector to provider
