+++ app/javascript/pages/mygfw/components/areas-table/selectors.js
@@ -1,9 +1,8 @@
 import { createStructuredSelector } from 'reselect';
 
-import { getTags } from 'components/map-menu/components/sections/my-gfw/selectors';
-import { getUserAreas } from 'providers/areas-provider/selectors';
+import { getUserAreas, getAreaTags } from 'providers/areas-provider/selectors';
 
 export const getAreasTableProps = createStructuredSelector({
   areas: getUserAreas,
-  tags: getTags
+  tags: getAreaTags
 });
