+++ app/javascript/providers/areas-provider/selectors.js
@@ -1,6 +1,8 @@
 import { createSelector, createStructuredSelector } from 'reselect';
 import sortBy from 'lodash/sortBy';
 import isEmpty from 'lodash/isEmpty';
+import flatMap from 'lodash/flatMap';
+import uniq from 'lodash/uniq';
 
 export const selectLoading = state => state.areas && state.areas.loading;
 export const selectLocation = state =>
@@ -15,6 +17,27 @@ export const getUserAreas = createSelector(
   [getAllAreas],
   areas => (areas && areas.filter(area => area.userArea)) || null
 );
+
+export const getActiveArea = createSelector(
+  [selectLocation, getAllAreas],
+  (location, areas) => {
+    if (isEmpty(areas)) return null;
+
+    return areas.find(a => a.id === location.adm0);
+  }
+);
+
+export const getAreaTags = createSelector([getUserAreas], areas => {
+  if (isEmpty(areas)) return null;
+
+  return sortBy(
+    uniq(flatMap(areas.map(area => area.tags))).map(t => ({
+      label: t.length > 15 ? t.substring(0, 15).concat('...') : t,
+      value: t
+    })),
+    'label'
+  );
+});
 
 export const getAreasProps = createStructuredSelector({
   areas: getAllAreas,
